aggiornamento funzioni wrapper

+++ lib/share_lib_client.h
@@ -83,12 +83,12 @@ private:
 };
 
 // calloc Wrapper
-void* Calloc(unsigned long int nmemb, unsigned long int size);
+void* Calloc(size_t nmemb, size_t size);
 // fprintf Wrapper
 void ShowErr(const char* str);
 // free Wrapper
-void Free(void* arg, int size=0);
+void Free(void* arg, size_t size = 0);
 // strcpy Wrapper
-void Strcpy(char* dest, unsigned int size, const char* src);
+void Strcpy(char* dest, size_t size, const char* src);
 // asprintf Wrapper
 int Asprintf(char*& buffer, const char* Format, ...);
